+++ _metadata.py
@@ -1,2 +1,2 @@
-__extension_version__ = "0.65.0"
+__extension_version__ = "0.66.0"
 __extension_name__ = "pytket-qiskit"